Reworked tile placement logic
We're no longer using standard setNextWorldTile as our direct position changing event. We're now going to be using a Task based event system to promptly handle this event with support of Animation & message. Code overall has been drastically reduced & more efficient.

Also started to reduce redundant code in the Player class (soon to be working on NPC class). Abstract and regular objects can be handled in Entity class (as long as it's not unethical to do so). This helps keep our main classes more tidy and organized and not having so much less-useful information being directly shown.